+++ src/systems/EntityAdapter/index.ts
@@ -79,3 +79,9 @@ export class EntityAdapter extends System {
 }
 
 export type { Component };
+export {
+	CanvasComponent,
+	PlayerEntityState,
+	SpaceState,
+	TurnCount,
+} from './components';
